+++ src/main/java/com/base/basesetup/dto/SummaryCostInvoiceDTO.java
@@ -1,5 +1,9 @@
 package com.base.basesetup.dto;
 
+import java.math.BigDecimal;
+
+import java.math.BigDecimal;
+
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
@@ -11,12 +15,12 @@ import lombok.NoArgsConstructor;
 @Builder
 public class SummaryCostInvoiceDTO {
 	private Long id;
-	private double billCurrTotChargeAmt;
-	private double billCurrActBillAmt;
-	private double billCurrNetAmt;
-	private double lcTotChargeAmt;
-	private double lcActBillAmt;
-	private double lcNetAmt;
+	private BigDecimal billCurrTotChargeAmt;
+	private BigDecimal billCurrActBillAmt;
+	private BigDecimal billCurrNetAmt;
+	private BigDecimal lcTotChargeAmt;
+	private BigDecimal lcActBillAmt;
+	private BigDecimal lcNetAmt;
 	private String roundOff;
-	private double lcGstInputAmt;
+	private BigDecimal lcGstInputAmt;
 }
